third_party/blink/web_tests/platform/linux-chrome/external/wpt/css/css-transitions/animations/transition-timing-function-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Unhandled rejection: assert_equals: Unexpected animation count expected 5 but got 6
Harness: the test ran to completion.
